requirements.txt
tensorflow>=2.13.0,<3.0.0
tensorflow-io>=0.32.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pyyaml>=6.0
